:boom: Refactor interrupt APIs
The previous interrupt APIs were born from libhal 1.0.0 where the policy
was to use header only libraries for all code. classes and structs make
building header only libraries quite easy along with the ability to
hide state in private sections. This is no longer needed now that we
are building all of our binaries and can use source files.

The goal of this change was to make the interrupt APIs a bit more
reasonable. Rather than `interrupt(irq_t).enable(handler)`, we simply
have `enable_interrupt(irq_t, handler)`. All of the APis were changed
to get rid of a lot of the unnecessary OOP we were doing to hide state.

This is a breaking API change that I believe needs to get in now before
there are too many platform libraries running about that will need to
be refactored later.

Also, negative IRQ types are now allowed for the cortex IRQ numbers.
This removes the need for platform developers to do the following when
creating their irq enumeration classes.

```
/// List of interrupt request numbers for this platform
enum class irq : std::uint16_t
{
  wdt = 16 + 0,
  timer0 = 16 + 1,
  timer1 = 16 + 2,
  // etc...
  max,
};
```

Now it can just be:

```
/// List of interrupt request numbers for this platform
enum class irq : std::uint16_t
{
  wdt = 0,
  timer0 = 1,
  timer1 = 2,
  // etc...
  max,
};
```

Which is more natural and ergonomic.